he web browser under Udacity course "Programming Languages"
The web browser under Udacity course "Programming Languages" Through the
help of free courseware "Programming Languages" under Prof. Westley
Weimer, built a simple web browser with the ability to parse and
understand HTML and basic Javascript. The web browser has lexer and
parser for HTML and Javascript for which PLY library is used. Abstract
Syntax trees are used to store the parse tree. HTML is interpreted with
the help of graphics library (provided online). Javascript is converted
to Python and the requisite operation is done by the graphics library.